refactor: 훅 폴더 정리
