+++ web/app/main/page.tsx
@@ -11,7 +11,7 @@ import { useCloverCount } from '@/hooks/queries/useCloverCount';
 import { useNotifications } from '@/hooks/queries/useNotifications';
 import TodayStatsCard from '@/components/main/TodayStatsCard';
 
-import { useAuthToken } from '@/hooks/useAuthToken';
+import { useAuthToken } from '@/hooks/user/useAuthToken';
 
 export default function Main() {
   const [displayNotifications, setDisplayNotifications] = useState<
